requirements.txt
asgiref==3.5.2
Django==3.2
django-crispy-forms==1.14.0
pytz==2022.1
sqlparse==0.4.2
django-mapbox-location-field==2.0.0
Pillow==9.2.0
django-location-field==2.7.0